React Router and Pagination added
