+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-router": "~7.2.0",
     "styled-components": "^6.1.18"
   },
   "devDependencies": {
