+++ src/App.module.css
@@ -1,25 +1,4 @@
 .block {
   display: grid;
-  justify-content: center;
-  align-items: center;
-}
-
-.title {
-  display: inline-flex;
-  align-items: center;
-}
-
-.logo {
-  height: 25px;
-  padding-right: 5px;
-}
-
-.error {
-  padding: 10px;
-  border-style: solid;
-  border-radius: 7px;
-  border-width: 1px;
-  border-color: darkgray;
-  background-color: rgba(230, 92, 92, 0.5);
 }
