+++ src/pages/TodosPage.module.css
@@ -1,0 +1,13 @@
+.block {
+  display: grid;
+}
+
+.error {
+  padding: 10px;
+  border-style: solid;
+  border-radius: 7px;
+  border-width: 1px;
+  border-color: darkgray;
+  background-color: rgba(230, 92, 92, 0.5);
+}
+
