+++ src/shared/Header.module.css
@@ -1,0 +1,34 @@
+.header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  padding: 10px;
+  box-sizing: border-box;
+  padding: 2rem 0;
+}
+
+.title {
+  margin: 0;
+}
+
+.logo {
+  height: 25px;
+  padding-right: 5px;
+}
+
+.nav {
+  padding: 1rem;
+}
+
+.active {
+  color: gray;
+  text-decoration: underline;
+  padding: .5rem;
+}
+
+.inactive {
+  color: black;
+  text-decoration: none;
+  padding: .5rem;
+}
